Moved new functions to make_grades_persistent
